Make Makefile more portable
To help make the Makefile more portable for different projects, the
PlaydateKit path is now extracted to PDKIT_ROOT, as it may not always be
the same as REPO_ROOT. PRODUCT_NAME was also introduced to make changing
the source paths for game source code easier, and PRODUCT uses
PRODUCT_NAME to its advantage.

Documentation on these variables are provided as comments here. In the
future, command line tools can help set these variables automatically
(such as with CMake) to automatically generate this Makefile.