Put -sdkpath argument before input and output arguments
It appears that, as of SDK v3.0.1, it has gotten a bit more picky
regarding the order of arguments. I noticed this caused my project
to stop building with `swift package pdc` (rather new template).

I'm not sure how this fares for older versions of the SDK, but I
can at least confirm this change works as expected on 3.0.1.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>